+++ test/logjam/event_test.clj
@@ -1,6 +1,7 @@
 (ns logjam.event-test
   (:require [clojure.set :as set]
             [clojure.spec.alpha :as s]
+            [clojure.test :refer [testing]]
             [clojure.test.check.clojure-test :refer [defspec]]
             [clojure.test.check.generators :as gen]
             [clojure.test.check.properties :as prop]
@@ -50,14 +51,38 @@
              (event/search (:levels framework) criteria events)))))
 
 (defspec test-search-loggers
-  (prop/for-all
-   [{:keys [levels]} (gen/elements frameworks)
-    loggers (s/gen :logjam.filter/loggers)
-    events (s/gen (s/coll-of :logjam/event))]
-   (let [opts {:filters {:loggers loggers}}
-         events-found (event/search levels opts events)]
-     (set/subset? (set (map :logger events-found))
-                  (set (map :logger events))))))
+  (testing "`:loggers` (legacy key)"
+    (prop/for-all
+     [{:keys [levels]} (gen/elements frameworks)
+      loggers (s/gen :logjam.filter/loggers)
+      events (s/gen (s/coll-of :logjam/event))]
+     (let [opts {:filters {:loggers loggers}}
+           events-found (event/search levels opts events)]
+       (set/subset? (set (map :logger events-found))
+                    (set (map :logger events)))))))
+
+(defspec test-search-loggers-allowlist
+  (testing "`:loggers-allowlist` (newer key)"
+    (prop/for-all
+     [{:keys [levels]} (gen/elements frameworks)
+      loggers (s/gen :logjam.filter/loggers-allowlist)
+      events (s/gen (s/coll-of :logjam/event))]
+     (let [opts {:filters {:loggers-allowlist loggers}}
+           events-found (event/search levels opts events)]
+       (set/subset? (set (map :logger events-found))
+                    (set (map :logger events)))))))
+
+(defspec test-search-loggers-blocklist
+  (testing "`:loggers-blocklist`"
+    (prop/for-all
+     [[events framework] (gen/let [framework (gen/elements frameworks)
+                                   events (gen/vector (test/event-gen framework) 10)]
+                           [events framework])]
+     (let [blocklist (->> events (take 5) (map :logger))
+           opts {:filters {:loggers-blocklist blocklist}}
+           events-found (event/search (:levels framework) opts events)]
+       (not-any? (set blocklist)
+                 (map :logger events-found))))))
 
 (defspec test-search-limit
   (prop/for-all