+++ test/logjam/specs.clj
@@ -29,6 +29,10 @@
 (s/def :logjam.filter/loggers
   (s/coll-of string? :kind set?))
 
+(s/def :logjam.filter/loggers-allowlist :logjam.filter/loggers)
+
+(s/def :logjam.filter/loggers-blocklist :logjam.filter/loggers)
+
 (s/def :logjam.filter/pattern string?)
 (s/def :logjam.filter/start-time pos-int?)
 
@@ -40,6 +44,8 @@
                    :logjam.filter/exceptions
                    :logjam.filter/level
                    :logjam.filter/loggers
+                   :logjam.filter/loggers-allowlist
+                   :logjam.filter/loggers-blocklist
                    :logjam.filter/pattern
                    :logjam.filter/start-time
                    :logjam.filter/threads]))
@@ -102,7 +108,8 @@
 (s/def :logjam.event/arguments (s/coll-of :logjam.event/argument :kind vector?))
 (s/def :logjam.event/id uuid?)
 (s/def :logjam.event/level simple-keyword?)
-(s/def :logjam.event/logger string?)
+(s/def :logjam.event/logger (s/with-gen string?
+                              #(s/gen (s/and string? (complement string/blank?)))))
 (s/def :logjam.event/mdc
   (s/with-gen map? ;; relaxed spec for Timbre
     #(s/gen (s/map-of string? string?)))) ;; strict gen for Logback